[4.1] Remove template name conversion (#76)
* refactor: form & form components now uses templateName instead of templateId for template handling

* refactor: improved error handling in form components

* refactor: further changes for template_name support

* refactor: moved funder import into its own Component
