+++ src/components/FormComponents/SubForm.jsx
@@ -8,7 +8,7 @@ import styles from '../assets/css/form.module.css';
 import NestedForm from '../Forms/NestedForm.jsx';
 import { useTranslation } from 'react-i18next';
 import { useController, useFormContext } from 'react-hook-form';
-import { fragmentEmpty } from '../../utils/utils.js';
+import { fragmentEmpty, getErrorMessage } from '../../utils/utils.js';
 import { parsePattern } from '../../utils/GeneratorUtils.js';
 import { GlobalContext } from '../context/Global.jsx';
 import CustomButton from '../Styled/CustomButton.jsx';
@@ -18,7 +18,7 @@ function SubForm({
   label,
   propName,
   tooltip,
-  templateId,
+  templateName,
   readonly = false,
 }) {
   const { t } = useTranslation();
@@ -40,15 +40,17 @@ function SubForm({
   }, [field.value])
 
   useEffect(() => {
-    if (!loadedTemplates[templateId]) {
-      service.getSchema(templateId).then((res) => {
+    if (!loadedTemplates[templateName]) {
+      service.getSchemaByName(templateName).then((res) => {
         setTemplate(res.data)
-        setLoadedTemplates({ ...loadedTemplates, [templateId]: res.data });
+        setLoadedTemplates({ ...loadedTemplates, [templateName]: res.data });
+      }).catch((error) => {
+        setError(getErrorMessage(error));
       });
     } else {
-      setTemplate(loadedTemplates[templateId]);
+      setTemplate(loadedTemplates[templateName]);
     }
-  }, [templateId])
+  }, [templateName])
 
 
   const handleSaveNestedForm = (data) => {
@@ -77,6 +79,7 @@ function SubForm({
   return (
     <div>
       <div className="form-group">
+        <span className={styles.errorMessage}>{error}</span>
         <div className={styles.label_form}>
           <strong className={styles.dot_label}></strong>
           <label data-tooltip-id={tooltipId}>{label}</label>
