+++ src/components/assets/css/form.module.css
@@ -211,7 +211,6 @@ input:checked + .switchSlider .switchOff {
   cursor: pointer;
 }
 
-
 .errorMessage {
   font-weight: bold;
   color: var(--rust);
