+++ src/utils/utils.js
@@ -23,7 +23,7 @@ export function except(data, excludedKeys) {
 }
 
 export function fragmentEmpty(data) {
-  let rest = except(data, ['id', 'schema_id', 'action']);
+  let rest = except(data, ['id', 'schema_id', 'template_name', 'action']);
   return Object.keys(rest).length === 0
 }
